Move system health check to local system checker and split it from the update checks. Do some opportunistic stream refactoring to remove warnings in LocalSystemChecker.
