+++ game-core/src/main/java/games/strategy/engine/auto/update/UpdateChecks.java
@@ -2,10 +2,6 @@ package games.strategy.engine.auto.update;
 
 import static games.strategy.engine.framework.ArgParser.CliProperties.TRIPLEA_GAME;
 
-import java.util.Collection;
-
-import games.strategy.debug.ClientLogger;
-import games.strategy.engine.auto.health.check.LocalSystemChecker;
 import games.strategy.engine.framework.map.download.MapDownloadController;
 
 /**
@@ -15,20 +11,9 @@ import games.strategy.engine.framework.map.download.MapDownloadController;
 public class UpdateChecks {
 
   public static void launch() {
-    new Thread(UpdateChecks::checkLocalSystem).start();
     new Thread(UpdateChecks::checkForUpdates).start();
   }
 
-  private static void checkLocalSystem() {
-    final LocalSystemChecker localSystemChecker = new LocalSystemChecker();
-    final Collection<Exception> exceptions = localSystemChecker.getExceptions();
-    if (!exceptions.isEmpty()) {
-      final String msg = String.format(
-          "Warning!! %d system checks failed. Some game features may not be available or may not work correctly.\n%s",
-          exceptions.size(), localSystemChecker.getStatusMessage());
-      ClientLogger.logError(msg);
-    }
-  }
 
   private static void checkForUpdates() {
     if (!UpdateCheckDecision.shouldRun()) {
